Allow to explicitly specify S3 bucket
Because of weird shenanigans of boto, rubberjack won't be able to deploy
application to a different account than the S3 bucket was created in.

Moreover, it has problems with deploying via a user (from the same
account that EB app is in) even if the S3 bucket is in the same AWS
account, but in a different region (probably some weird permission
issue). Nevertheless, I feel like this can be a useful option.